+++ widget/entry.go
@@ -229,18 +229,6 @@ func (e *Entry) FocusLost() {
 	})
 }
 
-// Focused returns whether or not this Entry has focus.
-//
-// Implements: fyne.Focusable
-//
-// Deprecated: this method will be removed as it is no longer required, widgets do not expose their focus state.
-func (e *Entry) Focused() bool {
-	e.propertyLock.RLock()
-	defer e.propertyLock.RUnlock()
-
-	return e.focused
-}
-
 // Hide hides the entry.
 //
 // Implements: fyne.Widget
@@ -883,9 +871,9 @@ func (e *Entry) textAlign() fyne.TextAlign {
 // textColor tells the rendering textProvider our color
 func (e *Entry) textColor() color.Color {
 	if e.Disabled() {
-		return theme.DisabledTextColor()
+		return theme.DisabledColor()
 	}
-	return theme.TextColor()
+	return theme.ForegroundColor()
 }
 
 // Obtains textual position from a given row and col
@@ -925,7 +913,7 @@ func (e *Entry) textWrap() fyne.TextWrap {
 }
 
 func (e *Entry) updateMousePointer(ev *fyne.PointEvent, rightClick bool) {
-	if !e.Focused() && !e.Disabled() {
+	if !e.focused && !e.Disabled() {
 		e.FocusGained()
 	}
 
@@ -1050,7 +1038,7 @@ func (r *entryRenderer) Refresh() {
 		r.line.FillColor = theme.FocusColor()
 	} else {
 		if r.entry.Disabled() {
-			r.line.FillColor = theme.DisabledTextColor()
+			r.line.FillColor = theme.DisabledColor()
 		} else {
 			r.line.FillColor = theme.ShadowColor()
 		}
@@ -1071,7 +1059,7 @@ func (r *entryRenderer) Refresh() {
 
 	if r.entry.Validator != nil {
 		if !r.entry.focused && r.entry.Text != "" && r.entry.validationError != nil {
-			r.line.FillColor = &color.NRGBA{0xf4, 0x43, 0x36, 0xff} // TODO: Should be current().ErrorColor() in the future
+			r.line.FillColor = theme.ErrorColor()
 		}
 		r.ensureValidationSetup()
 		r.entry.validationStatus.Refresh()
@@ -1476,7 +1464,7 @@ func (p *placeholderPresenter) textAlign() fyne.TextAlign {
 // textColor tells the rendering textProvider our color
 func (p *placeholderPresenter) textColor() color.Color {
 	if p.e.Disabled() {
-		return theme.DisabledTextColor()
+		return theme.DisabledColor()
 	}
 	return theme.PlaceHolderColor()
 }