+++ widget/slider.go
@@ -194,9 +194,9 @@ func (s *Slider) MinSize() fyne.Size {
 func (s *Slider) CreateRenderer() fyne.WidgetRenderer {
 	s.ExtendBaseWidget(s)
 	track := canvas.NewRectangle(theme.ShadowColor())
-	active := canvas.NewRectangle(theme.TextColor())
+	active := canvas.NewRectangle(theme.ForegroundColor())
 	thumb := &canvas.Circle{
-		FillColor:   theme.TextColor(),
+		FillColor:   theme.ForegroundColor(),
 		StrokeWidth: 0}
 
 	objects := []fyne.CanvasObject{track, active, thumb}
@@ -237,8 +237,8 @@ type sliderRenderer struct {
 // Refresh updates the widget state for drawing.
 func (s *sliderRenderer) Refresh() {
 	s.track.FillColor = theme.ShadowColor()
-	s.thumb.FillColor = theme.TextColor()
-	s.active.FillColor = theme.TextColor()
+	s.thumb.FillColor = theme.ForegroundColor()
+	s.active.FillColor = theme.ForegroundColor()
 
 	s.slider.clampValueToRange()
 	s.Layout(s.slider.Size())