+++ widget/tree_internal_test.go
@@ -193,6 +193,7 @@ func TestTree_Resize(t *testing.T) {
 	addTreePath(data, "B", "C")
 
 	tree.OpenBranch("B")
+	separatorThickness := theme.SeparatorThicknessSize()
 
 	width := templateMinSize.Width + indentation() + theme.IconInlineSize() + doublePadding + theme.Padding()
 	height := (fyne.Max(templateMinSize.Height, theme.IconInlineSize())+doublePadding)*3 + separatorThickness*2
@@ -219,6 +220,7 @@ func TestTree_Resize(t *testing.T) {
 }
 
 func TestTree_MinSize(t *testing.T) {
+	separatorThickness := theme.SeparatorThicknessSize()
 	t.Run("Default", func(t *testing.T) {
 		tree := &Tree{}
 		min := tree.MinSize()
@@ -404,6 +406,7 @@ func TestTree_ScrollToSelection(t *testing.T) {
 
 	a := getLeaf(t, tree, "A")
 	m := a.MinSize()
+	separatorThickness := theme.SeparatorThicknessSize()
 
 	// Make tree tall enough to display two nodes
 	tree.Resize(fyne.NewSize(m.Width, m.Height*2+separatorThickness))