+++ CHANGELOG.md
@@ -31,6 +31,9 @@ These changes likely break some apps, please read the
 
 * Removed deprecated types including:
   - dialog.FileIcon (now widget.FileIcon)
+  - widget.Radio (now widget.RadioGroup)
+  - widget.AccordionContainer (now widget.Accordion)
+  - layout.NewFixedGridLayout() (now layout.NewGridWrapLayout())
 
 ### Added